Updating save to the new Toml API
### What's done:
- new API of ktoml (0.2.7)
- changed Int to Long as Toml does not support Int type

+++ save-plugins/warn-plugin/src/commonMain/kotlin/org/cqfn/save/plugin/warn/WarnPlugin.kt
@@ -51,7 +51,7 @@ class WarnPlugin(
         return warnPluginConfig.wildCardInDirectoryMode?.let {
             handleTestFile(files.map { it.single() }.toList(), warnPluginConfig, generalConfig).asSequence()
         } ?: run {
-            files.chunked(warnPluginConfig.batchSize!!).flatMap { chunk ->
+            files.chunked(warnPluginConfig.batchSize!!.toInt()).flatMap { chunk ->
                 handleTestFile(chunk.map { it.single() }, warnPluginConfig, generalConfig)
             }
         }
